requirements.txt
numpy
tensorflow
h5py
absl-py